requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
alembic
psycopg2
pytest
requests
python-dotenv